requirements.txt
Django==3.2.7
djangorestframework==3.12.4
django-allauth==0.45.0
django-rest-auth==0.9.5
django-imagekit==4.0.2